backend/requirements.txt
smolagents[toolkit]
pandas
openpyxl
matplotlib
seaborn
python-dotenv
huggingface_hub
fastapi
uvicorn[standard]
python-multipart
